requirements.txt
asgiref==3.8.1
black==24.4.1
cfgv==3.4.0
click==8.1.7
distlib==0.3.8
Django==5.0
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.13.4
identify==2.5.36
inflection==0.5.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pre-commit==3.7.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
uritemplate==4.1.1
virtualenv==20.26.0
